Add logic to prevent E_NOTICE from list * Run phpcbf to clean up code * Add namespace
